+++ src/examplemod/java/commoble/looot/examplemod/LoootExampleMod.java
@@ -1,6 +1,6 @@
 package commoble.looot.examplemod;
 
-import net.minecraftforge.fml.common.Mod;
+import net.neoforged.fml.common.Mod;
 
 @Mod(LoootExampleMod.MODID)
 public class LoootExampleMod